SensorServer deleted, ISensorScene interface and getVisibleSceneObjects() added to Scene, SceneTest extended
