+++ SW/environment/src/hu/nik/project/environment/ISensorScene.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 /**
  * Created by zhodvogner on 2016. márc. 19.
  */
-public interface ISensorServer {
+public interface ISensorScene {
 
     public ArrayList<SceneObject> getVisibleSceneObjects(ScenePoint observerBase, int observerRotation, int viewAngle );
     public ArrayList<SceneObject> getVisibleSceneObjects(ScenePoint observerBase, int observerRotation, int viewAngle, int viewDistance );
